requirements.txt
gradio
requests
langchain
langchain-community
langchain-core
langchain-huggingface
langgraph
langchain-openai
duckduckgo-search
requests
huggingface-hub
langgraph
python-dotenv
wikipedia